move match-syntax logic
