+++ src/Doc/methods/transform/03-split.js
@@ -1,4 +1,4 @@
-const parseSyntax = require('../../syntax')
+const parseSyntax = require('../../../World/match-syntax')
 
 /** return a Document with three parts for every match
  * seperate everything before the word, as a new phrase
