+++ src/World/match-syntax/01-parseToken.js
